bridge_v2: run as a stateful host service instead of in docker
The bridge needs to be ONE process (it's stateful — sessions, the noid
map, mongo connection, Elko upstream all live in memory) AND survive
tableflip's fork+exec graceful upgrade. Docker can't host this:
PID 1 exit takes the container down with it, killing the tableflip
child along with the parent. systemd Type=simple has the same problem
(KillMode=control-group nukes the cgroup when the main PID exits).
The fix is a coordinated set of changes so one process serves every
port and survives a hot-restart cleanly under systemd Type=notify.

Changes:

  * Multi-listen: --listen now accumulates (stringSliceFlag in main.go).
    Bridge struct holds []net.Listener + []string; Run() spawns one
    Accept goroutine per listener; SnapshotAllWithTCP closes them all
    and walks /proc to force-close any tableflip-dup'd fds whose port
    matches any configured listen port. listenPort() → listenPorts()
    returns a set. SetListener → SetListeners. Lets a single process
    serve 1337 (C64) + 1986 (legacy QLink) + 2026 (bots) without
    splitting session state across processes.

  * sd_notify(MAINPID): after upg.Ready(), the new (forked) child
    sends READY=1\nMAINPID=<pid> via go-systemd's daemon.SdNotify.
    Combined with the unit's Type=notify + NotifyAccess=all (in the
    bridge_host ansible role), systemd reparents its main-PID
    tracking from the dying tableflip parent to the new child BEFORE
    the parent exits. Without this, when the parent exits cleanly
    post-Ready(), systemd sees its tracked main PID die and (with
    KillMode=control-group, the default) kills the entire cgroup —
    including the new child. Confirmed working: clean PID handoff
    in <1 second, listeners stay bound, no cgroup death.

  * Speaker / Appearing widened to *uint16. Elko emits noid=256
    (UNASSIGNED_NOID) as a sentinel before assignment; the previous
    *uint8 typing made elkoReader's json.Unmarshal blow up on
    OBJECTSPEAK_$ and APPEARING_$ broadcasts, dropping the message
    entirely. JSON-passthrough clients (sage, eliza) silently missed
    arrival/speech events; binary clients would have hit the same
    bug if a 256 ever leaked into those op handlers. Added noidU8()
    narrowing helper in utils.go that maps the 256 sentinel to
    GHOST_NOID (255) for the binary-protocol encode path; mirrors
    the pattern already in place for HabitatMod.Noid.

  * SnapshotAllWithTCP fast-skips JSON-passthrough sessions instead
    of waiting 10s per-session for the snapshotReq channel to time
    out. Those sessions have no TCP_REPAIR-worthy in-memory state to
    preserve and the JSON clients (bots, web) auto-reconnect via
    shouldReconnect. With three bots, a reload now completes in <2s
    instead of 30s+. C64 binary sessions still go through the full
    snapshot path and are TCP_REPAIR'd into the new process — that's
    what the whole zero-downtime story exists for.

Tests: stringSliceFlag accumulation/String, NewBridge defensive copy
of listenAddrs, listenPorts() for ipv4/ipv6/duplicates/garbage,
SetListeners round-trip, noidU8 nil/sentinel/in-range branches.

Confirmed end-to-end on the made: systemctl reload yields sub-second
PID handoff, listeners stay bound through the parent exit, MAINPID
goes from old to child cleanly, JSON bots reconnect immediately, no
"address already in use" or "Deactivated successfully" surprises.

+++ bridge_v2/bridge/schema.go
@@ -9,7 +9,12 @@ type ElkoMessage struct {
 	Adjustment           *uint8         `json:"adjustment,omitempty"`
 	AmountLo             *uint8         `json:"amount_lo,omitempty"`
 	AmountHi             *uint8         `json:"amount_hi,omitempty"`
-	Appearing            *uint8         `json:"appearing,omitempty"`
+	// *uint16 (not *uint8) because Elko uses UNASSIGNED_NOID (256) as a
+	// sentinel for an avatar that hasn't been assigned a region noid
+	// yet — the bridge's binary path narrows to GHOST_NOID at encode
+	// time, but for JSON-passthrough clients we'd silently drop the
+	// whole APPEARING_$ message if json.Unmarshal blew up here.
+	Appearing            *uint16        `json:"appearing,omitempty"`
 	ArgCount             *uint8         `json:"argCount,omitempty"`
 	AttackResult         *uint8         `json:"ATTACK_result,omitempty"`
 	AttackTarget         *uint8         `json:"ATTACK_target,omitempty"`
@@ -78,7 +83,10 @@ type ElkoMessage struct {
 	SeatId               *uint8         `json:"seat_id,omitempty"`
 	SfxNumber            *uint8         `json:"sfx_number,omitempty"`
 	Slot                 *uint8         `json:"slot,omitempty"`
-	Speaker              *uint8         `json:"speaker,omitempty"`
+	// *uint16 for the same UNASSIGNED_NOID sentinel reason as Appearing
+	// above. OBJECTSPEAK_$ broadcasts ("X has arrived") use speaker=256
+	// when the noid hasn't been assigned yet.
+	Speaker              *uint16        `json:"speaker,omitempty"`
 	SpraySprayee         *uint8         `json:"SPRAY_SPRAYEE,omitempty"`
 	SpraySuccess         *uint8         `json:"SPRAY_SUCCESS,omitempty"`
 	SprayCustomize0      *uint8         `json:"SPRAY_CUSTOMIZE_0,omitempty"`
